+++ packages/core/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./types";
 export { runChecks } from "./runner";
-export { getProfile, listChecks, listProfiles } from "./registry";
+export { getProfile, listChecks, listProfiles, resolveCheckId } from "./registry";